Fixed various compiler warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@619 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ distr/src/apps/org/jnode/apps/edit/Editor.java
@@ -148,7 +148,7 @@ public class Editor extends JFrame {
             public Object run() {
                 try {
                     FileInputStream fis = new FileInputStream(file);
-                    byte[] data = new byte[(int) fis.available()];
+                    byte[] data = new byte[fis.available()];
                     fis.read(data);
                     textArea.setText(new String(data));
                     fis.close();
